Chore: Move annotationservice back
